move RawResponse and MappedResponse into apollo-federation
