+++ apollo-federation/src/connectors/runtime.rs
@@ -5,3 +5,4 @@ pub mod http_json_transport;
 pub mod inputs;
 pub mod key;
 pub mod mapping;
+pub mod responses;
