+++ apollo-router/src/plugins/telemetry/config_new/connector/selectors.rs
@@ -1,5 +1,6 @@
 use apollo_federation::connectors::runtime::http_json_transport::TransportRequest;
 use apollo_federation::connectors::runtime::http_json_transport::TransportResponse;
+use apollo_federation::connectors::runtime::responses::MappedResponse;
 use derivative::Derivative;
 use opentelemetry::Array;
 use opentelemetry::StringValue;
@@ -9,7 +10,6 @@ use serde::Deserialize;
 use tower::BoxError;
 
 use crate::Context;
-use crate::plugins::connectors::handle_responses::MappedResponse;
 use crate::plugins::telemetry::config::AttributeValue;
 use crate::plugins::telemetry::config_new::Selector;
 use crate::plugins::telemetry::config_new::Stage;
@@ -310,6 +310,7 @@ mod tests {
     use apollo_federation::connectors::runtime::http_json_transport::TransportResponse;
     use apollo_federation::connectors::runtime::key::ResponseKey;
     use apollo_federation::connectors::runtime::mapping::Problem;
+    use apollo_federation::connectors::runtime::responses::MappedResponse;
     use http::HeaderValue;
     use http::StatusCode;
     use opentelemetry::Array;
@@ -320,7 +321,6 @@ mod tests {
     use super::ConnectorSource;
     use super::MappingProblems;
     use crate::Context;
-    use crate::plugins::connectors::handle_responses::MappedResponse;
     use crate::plugins::telemetry::config_new::Selector;
     use crate::plugins::telemetry::config_new::selectors::ResponseStatus;
     use crate::services::connector::request_service::Request;
